// src/types/types.ts
export type Game = 'Clash of Clans' | 'Clash Royale' | 'Brawl Stars';

export type GameDomains =
	| 'players'
	| 'clans'
	| 'clubs'
	| 'leagues'
	| 'locations'
	| 'goldpass'
	| 'esports'
	| 'labels'
	| 'cards'
	| 'clanwarleagues'
	| 'tournaments'
	| 'challenges'
	| 'globaltournaments'
	| 'rankings'
	| 'brawlers'
	| 'events';